+++ web/views/statuses.erb
@@ -144,7 +144,7 @@
   <% end %>
   <% if @statuses.empty? %>
     <tr>
-      <td colspan="6"></td>
+      <td colspan="8"></td>
     </tr>
   <% end %>
 </table>
